HW02 - Calculator
This is the code that I wrote for the second homework, which asks to create a "calculator" that accepts an initial input from the user of which operation they want to execute: add, subtract, multiply, divide, and alphabetize. Then, depending on which one they select, they are asked to provide more input that is specific to the operation. Error checking is performed to provide a message back to the user if incorrect input is given.